Phase 2 COMPLETE: All 31 temperature conversion errors fixed, build succeeds

+++ crates/core/src/core_types/atmospheric.rs
@@ -35,10 +35,10 @@ pub(crate) fn simulate_plume_rise(grid: &mut SimulationGrid, source_positions: &
             let source_temp = source_cell.temperature;
             let source_smoke = source_cell.smoke_particles;
 
-            if f32::from(source_temp) > *grid.ambient_temperature + 50.0 {
+            if source_temp > (*grid.ambient_temperature as f32) + 50.0 {
                 // Calculate plume rise velocity
-                let temp_excess = f32::from(source_temp) - *grid.ambient_temperature;
-                let buoyancy_vel = (2.0 * 9.81 * temp_excess / *grid.ambient_temperature).sqrt();
+                let temp_excess = source_temp - (*grid.ambient_temperature as f32);
+                let buoyancy_vel = (2.0 * 9.81 * temp_excess / (*grid.ambient_temperature as f32)).sqrt();
 
                 // Transport to cells above
                 let rise_distance = buoyancy_vel * dt;
